add stage checking; add depends checking

+++ PKGBUILD
@@ -105,6 +105,8 @@ pkgver() {
     echo ${ver:0:7}
 }
 
+PREFIX_TMP="${srcdir}/tmptmp"
+
 FORMAT_NAME='arm'
 FORMAT_MAGIC='\x7fELF\x01\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\x28\x00'
 FORMAT_MASK='\xff\xff\xff\xff\xff\xff\xff\x00\xff\xff\xff\xff\xff\xff\xff\xff\xfe\xff\xff\xff'
@@ -156,16 +158,32 @@ kali_rootfs_debootstrap() {
     sudo mount -o bind "${DN_APT_CACHE}" "${DN_ROOTFS_DEBIAN}/var/cache/apt/archives"
 
     echo "[DBG] debootstrap state 1"
-    # create the rootfs - not much to modify here, except maybe the hostname.
-    echo "[DBG] debootstrap --foreign --arch ${MACHINEARCH} kali '${DN_ROOTFS_DEBIAN}'  http://${INSTALL_MIRROR}/kali"
-    sudo debootstrap --foreign --arch ${MACHINEARCH} kali "${DN_ROOTFS_DEBIAN}" "http://${INSTALL_MIRROR}/kali"
+    if [[ -f "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE1" ]]; then
+        echo "[DBG] SKIP debootstrap state 1"
+
+    else
+        # create the rootfs - not much to modify here, except maybe the hostname.
+        echo "[DBG] debootstrap --foreign --arch ${MACHINEARCH} kali '${DN_ROOTFS_DEBIAN}'  http://${INSTALL_MIRROR}/kali"
+        sudo debootstrap --foreign --arch ${MACHINEARCH} kali "${DN_ROOTFS_DEBIAN}" "http://${INSTALL_MIRROR}/kali"
+        if [ "$?" = "0" ]; then
+            touch "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE1"
+        fi
+    fi
 
     if [ "${ISCROSS}" = "1" ]; then
         sudo cp /usr/bin/qemu-arm-static "${DN_ROOTFS_DEBIAN}/usr/bin/"
     fi
 
     echo "[DBG] debootstrap state 2"
-    sudo chroot "${DN_ROOTFS_DEBIAN}" /usr/bin/env -i LANG=C /debootstrap/debootstrap --second-stage
+    if [[ -f "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE2" ]]; then
+        echo "[DBG] SKIP debootstrap state 2"
+
+    else
+        sudo chroot "${DN_ROOTFS_DEBIAN}" /usr/bin/env -i LANG=C /debootstrap/debootstrap --second-stage
+        if [ "$?" = "0" ]; then
+            touch "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE2"
+        fi
+    fi
 
     echo "[DBG] debootstrap state 2.5"
     # Create sources.list
@@ -208,13 +226,6 @@ EOF
     export LC_ALL=C
     export DEBIAN_FRONTEND=noninteractive
 
-    #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/proc"
-    #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/dev/"
-    #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/dev/pts"
-    sudo mount -t proc proc "${DN_ROOTFS_DEBIAN}/proc"
-    sudo mount -o bind /dev/ "${DN_ROOTFS_DEBIAN}/dev/"
-    sudo mount -o bind /dev/pts "${DN_ROOTFS_DEBIAN}/dev/pts"
-
     cat << EOF > /tmp/deb
 console-common console-data/keymap/policy select Select keymap from full list
 console-common console-data/keymap/full select en-latin1-nodeadkeys
@@ -222,7 +233,18 @@ EOF
     sudo mv /tmp/deb "${DN_ROOTFS_DEBIAN}/debconf.set"
 
     echo "[DBG] debootstrap state 3"
-    cat << EOF > /tmp/ths
+    if [[ -f "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE3" ]]; then
+        echo "[DBG] SKIP debootstrap state 3"
+
+    else
+        #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/proc"
+        #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/dev/"
+        #sudo mkdir -p "${DN_ROOTFS_DEBIAN}/dev/pts"
+        sudo mount -t proc proc "${DN_ROOTFS_DEBIAN}/proc"
+        sudo mount -o bind /dev/ "${DN_ROOTFS_DEBIAN}/dev/"
+        sudo mount -o bind /dev/pts "${DN_ROOTFS_DEBIAN}/dev/pts"
+
+        cat << EOF > /tmp/ths
 #!/bin/bash
 export PATH=/bin:/sbin:/usr/bin:/usr/sbin:$PATH
 export DEBIAN_FRONTEND=noninteractive
@@ -233,13 +255,13 @@ echo -e "#!/bin/sh\nexit 101" > /usr/sbin/policy-rc.d
 chmod +x /usr/sbin/policy-rc.d
 
 apt-get update
-apt-get install locales-all
+apt-get --yes --force-yes install locales-all
 
 debconf-set-selections /debconf.set
 rm -f /debconf.set
 apt-get update
-apt-get -y install git-core binutils ca-certificates initramfs-tools uboot-mkimage
-apt-get -y install locales console-common less nano git
+apt-get --yes --force-yes install git-core binutils ca-certificates initramfs-tools uboot-mkimage
+apt-get --yes --force-yes install locales console-common less nano git
 echo "root:toor" | chpasswd
 sed -i -e 's/KERNEL\!=\"eth\*|/KERNEL\!=\"/' /lib/udev/rules.d/75-persistent-net-generator.rules
 rm -f /etc/udev/rules.d/70-persistent-net.rules
@@ -253,12 +275,15 @@ dpkg-divert --remove --rename /usr/sbin/invoke-rc.d
 
 rm -f /third-stage
 EOF
-    chmod +x /tmp/ths
-    sudo mv /tmp/ths "${DN_ROOTFS_DEBIAN}/third-stage"
+        chmod +x /tmp/ths
+        sudo mv /tmp/ths "${DN_ROOTFS_DEBIAN}/third-stage"
 
-    sudo chroot "${DN_ROOTFS_DEBIAN}" /third-stage
-
-    sudo umount "${DN_ROOTFS_DEBIAN}/var/cache/apt/archives"
+        sudo chroot "${DN_ROOTFS_DEBIAN}" /third-stage
+        if [ "$?" = "0" ]; then
+            touch "${PREFIX_TMP}_FLG_KALI_ROOTFS_STAGE3"
+        fi
+        sudo umount "${DN_ROOTFS_DEBIAN}/var/cache/apt/archives"
+    fi
 
     cat << EOF > /tmp/cln
 #!/bin/bash
@@ -326,6 +351,7 @@ EOF
     cp ${srcdir}/rpiwiggle-git/rpi-wiggle ${DN_ROOTFS_RPI2}/scripts/rpi-wiggle.sh
     chmod 755 ${DN_ROOTFS_RPI2}/scripts/rpi-wiggle.sh
 fi
+
 }
 
 rsync_and_verify() {
@@ -353,7 +379,6 @@ if [ 1 = 0 ]; then
 fi
 }
 
-
 # create a image file with two partitions: /boot/ and /
 kali_create_image() {
     PARAM_DN_ROOTFS_DEBIAN="$1"
@@ -361,18 +386,18 @@ kali_create_image() {
     PARAM_DN_ROOTFS_RPI2="$1"
     shift
 
-    FN_IMAGE="${srcdir}/${pkgname}-${pkgver}-${MACHINEARCH}.img"
     # Create the disk and partition it
-
     echo "Creating image file for ${pkgdesc}"
-    if [ ! -f "${FN_IMAGE}" ]; then
+    if [[ ! -f "${FN_IMAGE}" || ! -f "${PREFIX_TMP}_FLG_KALI_CREATE_IMAGE" ]]; then
         dd if=/dev/zero of=${FN_IMAGE} bs=1M count=${IMGCONTAINER_SIZE}
         parted ${FN_IMAGE} --script -- mklabel msdos
         #parted ${FN_IMAGE} --script -- mkpart primary fat32  0 64
         #parted ${FN_IMAGE} --script -- mkpart primary ext4  64 -1
         parted ${FN_IMAGE} --script -- mkpart primary fat32   2048s 264191s
         parted ${FN_IMAGE} --script -- mkpart primary ext4  264192s    100%
+        touch "${PREFIX_TMP}_FLG_KALI_CREATE_IMAGE"
     fi
+    #install_hardkernel_uboot ${FN_IMAGE}
 
     # Set the partition variables
     DEV_LOOP=$(sudo losetup -f --show ${FN_IMAGE})
@@ -463,12 +488,30 @@ my_setevn() {
     esac
     export MACHINEARCH="${MACHINE}"
 
+    export FN_IMAGE="${srcdir}/${pkgname}-${pkgver}-${MACHINEARCH}.img"
+
     export DN_TOOLCHAIN_UBOOT="${srcdir}/toolchains-uboot-${MACHINEARCH}"
     export DN_TOOLCHAIN_KERNEL="${srcdir}/toolchains-kernel-${MACHINEARCH}"
 
     DN_ROOTFS_RPI2="${srcdir}/rootfs-rpi2-${MACHINEARCH}"
     DN_BOOT="${DN_ROOTFS_RPI2}/boot"
     DN_ROOTFS_DEBIAN="${srcdir}/rootfs-kali-${MACHINEARCH}"
+
+    export ARCH=arm
+    if [ "${ISCROSS}" = "1" ]; then
+        if [ $(uname -m) = x86_64 ]; then
+            export DN_TOOLCHAIN_KERNEL="${srcdir}/tools-raspberrypi-git/"
+            export PATH=${DN_TOOLCHAIN_KERNEL}/arm-bcm2708/gcc-linaro-arm-linux-gnueabihf-raspbian-x64/bin/:$PATH
+            export CROSS_COMPILE=arm-linux-gnueabihf-
+        else
+            export DN_TOOLCHAIN_KERNEL="${srcdir}/tools-raspberrypi-git/"
+            export PATH=${DN_TOOLCHAIN_KERNEL}/arm-bcm2708/gcc-linaro-arm-linux-gnueabihf-raspbian/bin/:$PATH
+            export CROSS_COMPILE=arm-linux-gnueabihf-
+        fi
+    else
+        export CROSS_COMPILE=
+        unset CROSS_COMPILE
+    fi
 }
 
 prepare_rpi2_kernel () {
@@ -487,6 +530,7 @@ prepare_rpi2_kernel () {
 
 prepare() {
     my_setevn
+    #rm -f "${FN_IMAGE}" ${PREFIX_TMP}*
 
     rm -rf ${DN_BOOT}
     rm -rf ${DN_ROOTFS_RPI2}
